@@ -1,12 +1,16 @@
 package me.weishu.kernelsu.ui.util
 
+import android.content.ContentResolver
 import android.content.Context
+import android.net.Uri
 import android.os.Build
+import android.os.ParcelFileDescriptor
 import android.system.Os
 import com.topjohnwu.superuser.ShellUtils
 import me.weishu.kernelsu.Natives
 import me.weishu.kernelsu.ui.screen.getManagerVersion
 import java.io.File
+import java.io.FileOutputStream
 import java.io.FileWriter
 import java.io.PrintWriter
 import java.time.LocalDateTime
@@ -22,11 +26,15 @@ fun getBugreportFile(context: Context): File {
     val tombstonesFile = File(bugreportDir, "tombstones.tar.gz")
     val dropboxFile = File(bugreportDir, "dropbox.tar.gz")
     val pstoreFile = File(bugreportDir, "pstore.tar.gz")
+    // Xiaomi/Readmi devices have diag in /data/vendor/diag
     val diagFile = File(bugreportDir, "diag.tar.gz")
+    val opulsFile = File(bugreportDir, "opuls.tar.gz")
     val bootlogFile = File(bugreportDir, "bootlog.tar.gz")
     val mountsFile = File(bugreportDir, "mounts.txt")
     val fileSystemsFile = File(bugreportDir, "filesystems.txt")
-    val ksuFileTree = File(bugreportDir, "ksu_tree.txt")
+    val adbFileTree = File(bugreportDir, "adb_tree.txt")
+    val adbFileDetails = File(bugreportDir, "adb_details.txt")
+    val ksuFileSize = File(bugreportDir, "ksu_size.txt")
     val appListFile = File(bugreportDir, "packages.txt")
     val propFile = File(bugreportDir, "props.txt")
     val allowListFile = File(bugreportDir, "allowlist.bin")
@@ -34,19 +42,22 @@ fun getBugreportFile(context: Context): File {
     val bootConfig = File(bugreportDir, "boot_config.txt")
     val kernelConfig = File(bugreportDir, "defconfig.gz")
 
-    val shell = getRootShell()
+    val shell = getRootShell(true)
 
     shell.newJob().add("dmesg > ${dmesgFile.absolutePath}").exec()
     shell.newJob().add("logcat -d > ${logcatFile.absolutePath}").exec()
     shell.newJob().add("tar -czf ${tombstonesFile.absolutePath} -C /data/tombstones .").exec()
     shell.newJob().add("tar -czf ${dropboxFile.absolutePath} -C /data/system/dropbox .").exec()
     shell.newJob().add("tar -czf ${pstoreFile.absolutePath} -C /sys/fs/pstore .").exec()
-    shell.newJob().add("tar -czf ${diagFile.absolutePath} -C /data/vendor/diag .").exec()
+    shell.newJob().add("tar -czf ${diagFile.absolutePath} -C /data/vendor/diag . --exclude=./minidump.gz").exec()
+    shell.newJob().add("tar -czf ${opulsFile.absolutePath} -C /mnt/oplus/op2/media/log/boot_log/ .").exec()
     shell.newJob().add("tar -czf ${bootlogFile.absolutePath} -C /data/adb/ksu/log .").exec()
 
     shell.newJob().add("cat /proc/1/mountinfo > ${mountsFile.absolutePath}").exec()
     shell.newJob().add("cat /proc/filesystems > ${fileSystemsFile.absolutePath}").exec()
-    shell.newJob().add("ls -alRZ /data/adb > ${ksuFileTree.absolutePath}").exec()
+    shell.newJob().add("busybox tree /data/adb > ${adbFileTree.absolutePath}").exec()
+    shell.newJob().add("ls -alRZ /data/adb > ${adbFileDetails.absolutePath}").exec()
+    shell.newJob().add("du -sh /data/adb/ksu/* > ${ksuFileSize.absolutePath}").exec()
     shell.newJob().add("cp /data/system/packages.list ${appListFile.absolutePath}").exec()
     shell.newJob().add("getprop > ${propFile.absolutePath}").exec()
     shell.newJob().add("cp /data/adb/ksu/.allowlist ${allowListFile.absolutePath}").exec()
@@ -74,7 +85,7 @@ fun getBugreportFile(context: Context): File {
         val uname = Os.uname()
         pw.println("KernelRelease: ${uname.release}")
         pw.println("KernelVersion: ${uname.version}")
-        pw.println("Mahcine: ${uname.machine}")
+        pw.println("Machine: ${uname.machine}")
         pw.println("Nodename: ${uname.nodename}")
         pw.println("Sysname: ${uname.sysname}")
 
@@ -82,6 +93,8 @@ fun getBugreportFile(context: Context): File {
         pw.println("KernelSU: $ksuKernel")
         val safeMode = Natives.isSafeMode
         pw.println("SafeMode: $safeMode")
+        val lkmMode = Natives.isLkmMode
+        pw.println("LKM: $lkmMode")
     }
 
     // modules
@@ -99,3 +112,4 @@ fun getBugreportFile(context: Context): File {
 
     return targetFile
 }
+
