@@ -9,8 +9,8 @@ API_HASH = "d524b414d21f4d37f08684c1df41ac9c"
 
 
 BOT_TOKEN = os.environ.get("BOT_TOKEN")
-CHAT_ID = int(os.environ.get("CHAT_ID"))
-MESSAGE_THREAD_ID = int(os.environ.get("MESSAGE_THREAD_ID"))
+CHAT_ID = os.environ.get("CHAT_ID")
+MESSAGE_THREAD_ID = os.environ.get("MESSAGE_THREAD_ID")
 COMMIT_URL = os.environ.get("COMMIT_URL")
 COMMIT_MESSAGE = os.environ.get("COMMIT_MESSAGE")
 RUN_URL = os.environ.get("RUN_URL")
@@ -41,12 +41,15 @@ def get_caption():
 
 
 def check_environ():
+    global CHAT_ID, MESSAGE_THREAD_ID
     if BOT_TOKEN is None:
         print("[-] Invalid BOT_TOKEN")
         exit(1)
     if CHAT_ID is None:
         print("[-] Invalid CHAT_ID")
         exit(1)
+    else:
+        CHAT_ID = int(CHAT_ID)
     if COMMIT_URL is None:
         print("[-] Invalid COMMIT_URL")
         exit(1)
@@ -62,6 +65,11 @@ def check_environ():
     if VERSION is None:
         print("[-] Invalid VERSION")
         exit(1)
+    if MESSAGE_THREAD_ID is None:
+        print("[-] Invaild MESSAGE_THREAD_ID")
+        exit(1)
+    else:
+        MESSAGE_THREAD_ID = int(MESSAGE_THREAD_ID)
 
 
 async def main():
@@ -73,7 +81,9 @@ async def main():
         print("[-] No files to upload")
         exit(1)
     print("[+] Logging in Telegram with bot")
-    async with await TelegramClient(session=None, api_id=API_ID, api_hash=API_HASH).start(bot_token=BOT_TOKEN) as bot:
+    script_dir = os.path.dirname(os.path.abspath(sys.argv[0]))
+    session_dir = os.path.join(script_dir, "ksubot")
+    async with await TelegramClient(session=session_dir, api_id=API_ID, api_hash=API_HASH).start(bot_token=BOT_TOKEN) as bot:
         caption = [""] * len(files)
         caption[-1] = get_caption()
         print("[+] Caption: ")
@@ -83,7 +93,6 @@ async def main():
         print("[+] Sending")
         await bot.send_file(entity=CHAT_ID, file=files, caption=caption, reply_to=MESSAGE_THREAD_ID, parse_mode="markdown")
         print("[+] Done!")
-        await bot.log_out()
 
 if __name__ == "__main__":
     try:
