@@ -1,7 +1,4 @@
 use anyhow::{Ok, Result};
-
-#[cfg(unix)]
-use anyhow::ensure;
 use getopts::Options;
 use std::env;
 #[cfg(unix)]
@@ -14,24 +11,17 @@ use crate::{
     utils::{self, umask},
 };
 
-pub const KERNEL_SU_OPTION: u32 = 0xDEAD_BEEF;
-
-const CMD_GRANT_ROOT: u64 = 0;
-// const CMD_BECOME_MANAGER: u64 = 1;
-const CMD_GET_VERSION: u64 = 2;
-// const CMD_ALLOW_SU: u64 = 3;
-// const CMD_DENY_SU: u64 = 4;
-// const CMD_GET_ALLOW_LIST: u64 = 5;
-// const CMD_GET_DENY_LIST: u64 = 6;
-const CMD_REPORT_EVENT: u64 = 7;
-pub const CMD_SET_SEPOLICY: u64 = 8;
-pub const CMD_CHECK_SAFEMODE: u64 = 9;
-
-const EVENT_POST_FS_DATA: u64 = 1;
-const EVENT_BOOT_COMPLETED: u64 = 2;
+#[cfg(any(target_os = "linux", target_os = "android"))]
+use rustix::{
+    process::getuid,
+    thread::{set_thread_res_gid, set_thread_res_uid, Gid, Uid},
+};
 
 #[cfg(any(target_os = "linux", target_os = "android"))]
-pub fn grant_root() -> Result<()> {
+pub fn grant_root(global_mnt: bool) -> Result<()> {
+    const KERNEL_SU_OPTION: u32 = 0xDEAD_BEEF;
+    const CMD_GRANT_ROOT: u64 = 0;
+
     let mut result: u32 = 0;
     unsafe {
         #[allow(clippy::cast_possible_wrap)]
@@ -44,12 +34,25 @@ pub fn grant_root() -> Result<()> {
         );
     }
 
-    ensure!(result == KERNEL_SU_OPTION, "grant root failed");
-    Err(std::process::Command::new("sh").exec().into())
+    anyhow::ensure!(result == KERNEL_SU_OPTION, "grant root failed");
+    let mut command = std::process::Command::new("sh");
+    let command = unsafe {
+        command.pre_exec(move || {
+            if global_mnt {
+                let _ = utils::switch_mnt_ns(1);
+                let _ = utils::unshare_mnt_ns();
+            }
+            std::result::Result::Ok(())
+        })
+    };
+    // add /data/adb/ksu/bin to PATH
+    #[cfg(any(target_os = "linux", target_os = "android"))]
+    add_path_to_env(defs::BINARY_DIR)?;
+    Err(command.exec().into())
 }
 
 #[cfg(not(any(target_os = "linux", target_os = "android")))]
-pub fn grant_root() -> Result<()> {
+pub fn grant_root(_global_mnt: bool) -> Result<()> {
     unimplemented!("grant_root is only available on android");
 }
 
@@ -58,23 +61,34 @@ fn print_usage(program: &str, opts: Options) {
     print!("{}", opts.usage(&brief));
 }
 
-fn set_identity(uid: u32, gid: u32) {
+fn set_identity(uid: u32, gid: u32, groups: &[u32]) {
     #[cfg(any(target_os = "linux", target_os = "android"))]
-    unsafe {
-        libc::seteuid(uid);
-        libc::setresgid(gid, gid, gid);
-        libc::setresuid(uid, uid, uid);
+    {
+        rustix::process::set_groups(
+            groups
+                .iter()
+                .map(|g| unsafe { Gid::from_raw(*g) })
+                .collect::<Vec<_>>()
+                .as_ref(),
+        )
+        .ok();
+        let gid = unsafe { Gid::from_raw(gid) };
+        let uid = unsafe { Uid::from_raw(uid) };
+        set_thread_res_gid(gid, gid, gid).ok();
+        set_thread_res_uid(uid, uid, uid).ok();
     }
 }
 
-#[cfg(not(unix))]
+#[cfg(not(any(target_os = "linux", target_os = "android")))]
 pub fn root_shell() -> Result<()> {
     unimplemented!()
 }
 
-#[cfg(unix)]
+#[cfg(any(target_os = "linux", target_os = "android"))]
 pub fn root_shell() -> Result<()> {
     // we are root now, this was set in kernel!
+
+    use anyhow::anyhow;
     let env_args: Vec<String> = std::env::args().collect();
     let program = env_args[0].clone();
     let args = env_args
@@ -116,6 +130,13 @@ pub fn root_shell() -> Result<()> {
         "M",
         "mount-master",
         "force run in the global mount namespace",
+    );
+    opts.optopt("g", "group", "Specify the primary group", "GROUP");
+    opts.optmulti(
+        "G",
+        "supp-group",
+        "Specify a supplementary group. The first specified supplementary group is also used as a primary group if the option -g is not specified.",
+        "GROUP",
     );
 
     // Replace -cn with -z, -mm with -M for supporting getopt_long
@@ -161,6 +182,23 @@ pub fn root_shell() -> Result<()> {
     let preserve_env = matches.opt_present("p");
     let mount_master = matches.opt_present("M");
 
+    let groups = matches
+        .opt_strs("G")
+        .into_iter()
+        .map(|g| g.parse::<u32>().map_err(|_| anyhow!("Invalid GID: {}", g)))
+        .collect::<Result<Vec<_>, _>>()?;
+
+    // if -g provided, use it.
+    let mut gid = matches
+        .opt_str("g")
+        .map(|g| g.parse::<u32>().map_err(|_| anyhow!("Invalid GID: {}", g)))
+        .transpose()?;
+
+    // otherwise, use the first gid of groups.
+    if gid.is_none() && !groups.is_empty() {
+        gid = Some(groups[0]);
+    }
+
     // we've make sure that -c is the last option and it already contains the whole command, no need to construct it again
     let args = matches
         .opt_str("c")
@@ -174,13 +212,12 @@ pub fn root_shell() -> Result<()> {
     }
 
     // use current uid if no user specified, these has been done in kernel!
-    let mut uid = unsafe { libc::getuid() };
-    let gid = unsafe { libc::getgid() };
+    let mut uid = getuid().as_raw();
     if free_idx < matches.free.len() {
         let name = &matches.free[free_idx];
         uid = unsafe {
             #[cfg(target_arch = "aarch64")]
-            let pw = libc::getpwnam(name.as_ptr() as *const u8).as_ref();
+            let pw = libc::getpwnam(name.as_ptr()).as_ref();
             #[cfg(target_arch = "x86_64")]
             let pw = libc::getpwnam(name.as_ptr() as *const i8).as_ref();
 
@@ -191,6 +228,8 @@ pub fn root_shell() -> Result<()> {
         }
     }
 
+    // if there is no gid provided, use uid.
+    let gid = gid.unwrap_or(uid);
     // https://github.com/topjohnwu/Magisk/blob/master/native/src/su/su_daemon.cpp#L408
     let arg0 = if is_login { "-" } else { &shell };
 
@@ -241,7 +280,7 @@ pub fn root_shell() -> Result<()> {
                 let _ = utils::unshare_mnt_ns();
             }
 
-            set_identity(uid, gid);
+            set_identity(uid, gid, &groups);
 
             std::result::Result::Ok(())
         })
@@ -259,54 +298,4 @@ fn add_path_to_env(path: &str) -> Result<()> {
     let new_path_env = env::join_paths(paths)?;
     env::set_var("PATH", new_path_env);
     Ok(())
-}
-
-pub fn get_version() -> i32 {
-    let mut result: i32 = 0;
-    #[cfg(any(target_os = "linux", target_os = "android"))]
-    unsafe {
-        #[allow(clippy::cast_possible_wrap)]
-        libc::prctl(
-            KERNEL_SU_OPTION as i32, // supposed to overflow
-            CMD_GET_VERSION,
-            std::ptr::addr_of_mut!(result).cast::<libc::c_void>(),
-        );
-    }
-    result
-}
-
-fn report_event(event: u64) {
-    #[cfg(any(target_os = "linux", target_os = "android"))]
-    unsafe {
-        #[allow(clippy::cast_possible_wrap)]
-        libc::prctl(
-            KERNEL_SU_OPTION as i32, // supposed to overflow
-            CMD_REPORT_EVENT,
-            event,
-        );
-    }
-}
-
-pub fn check_kernel_safemode() -> bool {
-    let mut result: i32 = 0;
-    #[cfg(any(target_os = "linux", target_os = "android"))]
-    unsafe {
-        #[allow(clippy::cast_possible_wrap)]
-        libc::prctl(
-            KERNEL_SU_OPTION as i32, // supposed to overflow
-            CMD_CHECK_SAFEMODE,
-            0,
-            0,
-            std::ptr::addr_of_mut!(result).cast::<libc::c_void>(),
-        );
-    }
-    result == KERNEL_SU_OPTION as i32
-}
-
-pub fn report_post_fs_data() {
-    report_event(EVENT_POST_FS_DATA);
-}
-
-pub fn report_boot_complete() {
-    report_event(EVENT_BOOT_COMPLETED);
 }
